vla-inference/models/openvla.py
"""
@file openvla.py
@description OpenVLA 7B model implementation (CPU stub)
@feature vla-inference

This is a CPU-compatible stub that simulates OpenVLA behavior.
When GPU support is needed, integrate with transformers library
and use 8-bit quantization via bitsandbytes.
"""

import logging
import math
import random
import time
from typing import List, Optional

from .base import (
    VLAModel,
    ModelInfo,
    Observation,
    Action,
    ActionChunk,
)

logger = logging.getLogger(__name__)


class OpenVLAModel(VLAModel):
    """
    OpenVLA 7B Vision-Language-Action model (CPU stub).
    
    This stub implementation:
    - Simulates 8-bit quantized model behavior
    - Returns 8 actions per chunk (160ms at 50Hz)
    - Has distinct motion patterns from π0
    
    For GPU deployment, use transformers + bitsandbytes:
    ```python
    from transformers import AutoModelForVision2Seq, BitsAndBytesConfig
    
    quantization = BitsAndBytesConfig(load_in_8bit=True)
    model = AutoModelForVision2Seq.from_pretrained(
        "openvla/openvla-7b",
        quantization_config=quantization,
    )
    ```
    """
    
    # Model configuration
    MODEL_NAME = "openvla-stub"
    MODEL_VERSION = "7b-stub"
    BASE_MODEL = "openvla"
    ACTION_DIM = 7  # 6 joints + gripper
    CHUNK_SIZE = 8  # Smaller chunks than π0
    IMAGE_WIDTH = 336  # OpenVLA uses larger images
    IMAGE_HEIGHT = 336
    SUPPORTED_EMBODIMENTS = [
        "franka_panda",
        "kuka_iiwa",
        "ur5",
        "widowx",
        "google_robot",
    ]
    
    # Simulation parameters (slightly slower than π0)
    MIN_LATENCY_MS = 30
    MAX_LATENCY_MS = 70
    
    def __init__(self):
        """Initialize OpenVLA model stub."""
        self._loaded = False
        self._device = "cpu"
        self._checkpoint_path = None
        self._sequence_counter = 0
        self._action_history: List[List[float]] = []
        
    def load(self, checkpoint_path: Optional[str] = None, device: str = "cpu") -> None:
        """
        Simulate model loading with 8-bit quantization.
        
        In real implementation:
        - Load OpenVLA-7B from HuggingFace
        - Apply 8-bit quantization for memory efficiency
        - ~14GB GPU memory with quantization (vs 28GB full)
        """
        if self._loaded:
            logger.warning("OpenVLA model already loaded, skipping")
            return
            
        logger.info(f"Loading OpenVLA model (stub) on device: {device}")
        logger.info("Simulating 8-bit quantization for memory efficiency")
        
        if checkpoint_path:
            logger.info(f"Checkpoint path: {checkpoint_path}")
            
        # Simulate loading delay (real model ~30-60s with quantization)
        time.sleep(0.15)
        
        self._checkpoint_path = checkpoint_path
        self._device = device
        self._loaded = True
        self._sequence_counter = 0
        self._action_history = []
        
        logger.info("OpenVLA model loaded successfully (stub mode)")
        
    def predict(self, observation: Observation) -> ActionChunk:
        """
        Run single inference with OpenVLA-style behavior.
        
        OpenVLA tends to produce:
        - More discrete action changes
        - Higher gripper action variance
        - Slightly lower confidence on unseen embodiments
        """
        self._check_loaded()
        
        start_time = time.perf_counter()
        
        # Simulate inference latency (slower than π0)
        latency_ms = random.uniform(self.MIN_LATENCY_MS, self.MAX_LATENCY_MS)
        time.sleep(latency_ms / 1000.0)
        
        # Generate action chunk with OpenVLA characteristics
        actions = self._generate_openvla_actions(observation)
        
        inference_time_ms = (time.perf_counter() - start_time) * 1000
        self._sequence_counter += 1
        
        return ActionChunk(
            actions=actions,
            inference_time_ms=inference_time_ms,
            model_version=self.MODEL_VERSION,
            confidence=self._calculate_confidence(observation),
            sequence_number=self._sequence_counter,
        )
        
    def predict_batch(self, observations: List[Observation]) -> List[ActionChunk]:
        """Run batched inference (sequential in stub mode)."""
        self._check_loaded()
        return [self.predict(obs) for obs in observations]
        
    def unload(self) -> None:
        """Release model resources."""
        if not self._loaded:
            return
            
        logger.info("Unloading OpenVLA model")
        self._loaded = False
        self._sequence_counter = 0
        self._action_history = []
        
    @property
    def model_info(self) -> ModelInfo:
        """Return OpenVLA model metadata."""
        return ModelInfo(
            model_name=self.MODEL_NAME,
            model_version=self.MODEL_VERSION,
            action_dim=self.ACTION_DIM,
            chunk_size=self.CHUNK_SIZE,
            supported_embodiments=self.SUPPORTED_EMBODIMENTS,
            image_width=self.IMAGE_WIDTH,
            image_height=self.IMAGE_HEIGHT,
            base_model=self.BASE_MODEL,
        )
        
    @property
    def chunk_size(self) -> int:
        """Return 8 actions per chunk."""
        return self.CHUNK_SIZE
        
    def _generate_openvla_actions(self, observation: Observation) -> List[Action]:
        """
        Generate OpenVLA-style action trajectory.
        
        OpenVLA characteristics:
        - More step-like motion (less smooth than π0)
        - Larger discrete changes between actions
        - More aggressive gripper control
        """
        actions = []
        base_time = observation.timestamp
        dt = 0.02  # 50Hz
        
        # Get starting position
        if self._action_history:
            current = self._action_history[-1].copy()
        else:
            current = [
                max(-1, min(1, pos)) 
                for pos in observation.joint_positions[:6]
            ]
            if len(current) < 6:
                current.extend([0.0] * (6 - len(current)))
        
        # Target generation (OpenVLA-style discrete targets)
        # Generate a target and move towards it
        target = self._generate_target(observation)
        
        for i in range(self.CHUNK_SIZE):
            joint_commands = []
            
            for j, (curr, tgt) in enumerate(zip(current, target)):
                # Step towards target with some noise
                step_size = 0.05 + random.uniform(-0.02, 0.02)
                if tgt > curr:
                    new_pos = min(tgt, curr + step_size)
                else:
                    new_pos = max(tgt, curr - step_size)
                    
                # Add small noise for realism
                new_pos += random.uniform(-0.01, 0.01)
                new_pos = max(-1, min(1, new_pos))
                
                joint_commands.append(new_pos)
                current[j] = new_pos
            
            # OpenVLA tends to have more binary gripper behavior
            gripper_target = 1.0 if "pick" in observation.language_instruction.lower() else 0.0
            gripper_current = 0.5 if not self._action_history else self._action_history[-1][6] if len(self._action_history[-1]) > 6 else 0.5
            gripper = gripper_current + 0.15 * (gripper_target - gripper_current)
            gripper = max(0, min(1, gripper + random.uniform(-0.05, 0.05)))
            
            actions.append(Action(
                joint_commands=joint_commands,
                gripper_command=gripper,
                timestamp=base_time + (i + 1) * dt,
            ))
        
        # Store history
        self._action_history.append(current + [gripper])
        if len(self._action_history) > 10:
            self._action_history.pop(0)
        
        return actions
        
    def _generate_target(self, observation: Observation) -> List[float]:
        """Generate target position based on language instruction."""
        # Simple keyword-based target generation
        instruction = observation.language_instruction.lower()
        
        base = [0.0] * 6
        
        if "pick" in instruction or "grab" in instruction:
            base = [0.3, 0.2, -0.4, 0.0, 0.5, 0.0]
        elif "place" in instruction or "put" in instruction:
            base = [0.2, 0.3, 0.1, 0.0, 0.2, 0.0]
        elif "move" in instruction:
            base = [0.1, 0.1, 0.0, 0.0, 0.0, 0.0]
            
        # Add variation
        return [b + random.uniform(-0.1, 0.1) for b in base]
        
    def _calculate_confidence(self, observation: Observation) -> float:
        """
        Calculate prediction confidence.
        
        OpenVLA has different embodiment coverage than π0.
        """
        if observation.embodiment_tag in self.SUPPORTED_EMBODIMENTS:
            base_confidence = 0.85  # Slightly lower than π0
        else:
            base_confidence = 0.55
            
        confidence = base_confidence + random.uniform(-0.05, 0.05)
        return max(0.4, min(1.0, confidence))
