vla-inference/requirements.txt
# VLA Inference Server Dependencies

# gRPC
grpcio>=1.60.0
grpcio-tools>=1.60.0

# Async support
aiohttp>=3.9.0

# GPU monitoring (optional - gracefully handles if unavailable)
pynvml>=11.5.0

# Configuration
python-dotenv>=1.0.0

# Metrics
prometheus-client>=0.20.0

# Numerical computation (for stub models)
numpy>=1.24.0

# Image validation (for stub models)
Pillow>=10.0.0

# ML frameworks (commented out - uncomment for GPU deployment)
# torch>=2.2.0
# transformers>=4.40.0
# accelerate>=0.28.0
# bitsandbytes>=0.43.0  # For 8-bit quantization
# openpi>=0.1.0  # For π0.6 model (when available)

# Development
pytest>=7.4.0
pytest-asyncio>=0.23.0
